LPS-91288 Rename
